requirements.txt
# Zilliz/Milvus相关依赖
pymilvus>=2.3.0

# 环境变量管理
python-dotenv>=1.0.0

# 基础数据处理
numpy>=1.21.0

# HTTP请求
requests>=2.25.0

# 必需依赖
# UUID生成（Python内置）
# datetime处理（Python内置）
# typing支持（Python内置）

# 可选依赖（根据需要安装）
# pandas>=1.3.0
# transformers>=4.20.0
# torch>=1.12.0
# sentence-transformers>=2.0.0
# loguru>=0.6.0
# tqdm>=4.60.0